Se implemento Json web token, y se configuraron endpoints para funcionar con el

+++ NAOM-api/src/main/java/org/generation/naomdb/service/OrdenesService.java
@@ -1,6 +1,5 @@
 package org.generation.naomdb.service;
 
-import org.aspectj.weaver.ast.Or;
 import org.generation.naomdb.exception.OrdenNotFound;
 import org.generation.naomdb.model.Ordenes;
 import org.generation.naomdb.model.Producto;
@@ -9,6 +8,7 @@ import org.generation.naomdb.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.ServletException;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
@@ -26,37 +26,32 @@ public class OrdenesService {
         this.usuarioRepository = usuarioRepository;
     }
 
-    public List<Ordenes> getOrdenesFromUsuario(Long id) throws OrdenNotFound {
-        Optional<List<Ordenes>> ordenesByUser = ordenesRepository.findOrdenesByUser(id);
-        if (ordenesByUser.isPresent()) {
-            return ordenesByUser.get();
-        }
-        throw new OrdenNotFound("El usuario con id " + id + " no tiene ordenes");
-    }
-
-    public Ordenes deleteOrden(Long id) throws OrdenNotFound {
+    public Ordenes deleteOrden(String correo, Long id) throws OrdenNotFound, ServletException {
         Optional<Ordenes> ordenesOpt = ordenesRepository.findById(id);
         if (ordenesOpt.isPresent()) {
             Ordenes ordenes = ordenesOpt.get();
-            ordenesRepository.deleteById(id);
-            return ordenes;
+            if(ordenes.getUsuario().getCorreo() == correo){
+                ordenesRepository.deleteById(id);
+                return ordenes;
+            }
+            throw new ServletException("You dont own this order");
         }
         throw new OrdenNotFound("Orden con el id " + id + " no se encuentra");
     }
 
-    public Ordenes updateOrden(Long id, int cantidad, BigDecimal totalOrden, List<Producto> productos) throws OrdenNotFound {
+    public Ordenes updateOrden(String correo,Long id, int cantidad, BigDecimal totalOrden, List<Producto> productos) throws OrdenNotFound, ServletException {
         Optional<Ordenes> ordenes = ordenesRepository.findById(id);
         if (ordenes.isPresent()) {
             Ordenes ord = ordenes.get();
-            if (cantidad != 0) ord.setCantidad(cantidad);
-            if (totalOrden != null) ord.setTotalOrden(totalOrden);
-            if (productos != null) ord.setProductos(productos);
-            ordenesRepository.save(ord);
-            return ord;
+            if(ord.getUsuario().getCorreo() == correo){
+                if (cantidad != 0) ord.setCantidad(cantidad);
+                if (totalOrden != null) ord.setTotalOrden(totalOrden);
+                if (productos != null) ord.setProductos(productos);
+                ordenesRepository.save(ord);
+                return ord;
+            }
+            throw new ServletException("You dont own this order");
         }
         throw new OrdenNotFound("Orden con el id " + id + " no se encuentra");
-
     }
-
-
 }
